+++ lib/dictd_reader.dart
@@ -1,3 +1,4 @@
 library dictd_reader;
 
 export 'src/dictd_reader.dart';
+export 'src/source.dart';